+++ Backend/src/main/java/br/com/java/pcbuild/repositories/CpuCoolerSocketCompatibilityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CpuCoolerSocketCompatibilityRepository extends JpaRepository<CpuCoolerSocketCompatibility, Integer> {
+public interface CpuCoolerSocketCompatibilityRepository extends JpaRepository<CpuCoolerSocketCompatibility, Long> {
