+++ Backend/src/main/java/br/com/java/pcbuild/repositories/GpuPerformanceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GpuPerformanceRepository extends JpaRepository<GpuPerformance, Integer> {
+public interface GpuPerformanceRepository extends JpaRepository<GpuPerformance, Long> {
